third_party/blink/web_tests/fast/css/invalidation/visited-pseudo-expected.txt

Use descendant invalidation set for :visited pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.computedStyleIncludingVisitedInfo(match).backgroundColor is red
PASS internals.updateStyleAndReturnAffectedElementCount() is 5
PASS internals.computedStyleIncludingVisitedInfo(match).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
